+++ src/solver.h
@@ -5,16 +5,16 @@
 namespace miso {
 
 void solve_min_isocline(
-        Eigen::Vector3d& min_isocline_direction,
-        double& min_isocline_length,
-        const Eigen::MatrixXd& v,
+        Eigen::Vector3f& min_isocline_direction,
+        float& min_isocline_length,
+        const Eigen::MatrixXf& v,
         const Eigen::MatrixXi& f,
-        const Eigen::MatrixXd& n,
-        const double min_temperature = 0.0001,
-        const double alpha = 0.9,
+        const Eigen::MatrixXf& n,
+        const float min_temperature = 0.0001f,
+        const float alpha = 0.9f,
         const uint32_t max_iterations = 100,
         const uint32_t max_inner_iterations = 100,
-        const double neighbor_stddev = 0.1,
+        const float neighbor_stddev = 0.1f,
         const bool verbose = false);
 
 }
